@@ -3,23 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.11",
+    "@fortawesome/fontawesome-svg-core": "^1.2.32",
+    "@fortawesome/free-solid-svg-icons": "^5.15.1",
+    "@fortawesome/react-fontawesome": "^0.1.14",
     "ace-builds": "^1.4.12",
-    "axios": "^0.20.0",
-    "bootstrap": "^4.5.2",
+    "axios": "^0.21.1",
+    "bootstrap": "^4.5.3",
     "katex": "^0.12.0",
     "lodash": "^4.17.20",
-    "moment": "^2.27.0",
-    "react": "^16.13.1",
-    "react-ace": "^9.1.3",
-    "react-dom": "^16.13.1",
-    "react-resize-detector": "^5.0.7",
+    "moment": "^2.29.1",
+    "react": "^16.14.0",
+    "react-ace": "^9.2.1",
+    "react-dom": "^16.14.0",
+    "react-resize-detector": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
+    "react-scripts": "4.0.1",
     "react-split": "^2.0.9",
-    "reactstrap": "^8.5.1"
+    "reactstrap": "^8.8.1"
+  },
+  "devDependencies": {
+    "typescript": "^4.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
