@@ -1,30 +1,33 @@
 {
   "name": "wkspace",
   "version": "1.0.0",
-  "description": "",
+  "description": "Competitive programming workspace in the cloud.",
   "private": true,
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "node index.js",
     "mongo": "mkdir .mongodb; mongod --dbpath .mongodb",
-    "dev": "concurrently \"npm run mongo\" \"nodemon index.js\" \"cd client && npm start\"",
-    "heroku-postbuild": "cd client && npm install && npm run build"
+    "dev": "concurrently \"npm run mongo\" \"sleep 3 && nodemon index.js\" \"cd client && npm start\"",
+    "heroku-postbuild": "cd client && npm install && npm run build",
+    "format": "prettier --write ."
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.20.0",
-    "cheerio": "^1.0.0-rc.3",
+    "axios": "^0.21.1",
+    "cheerio": "^1.0.0-rc.5",
     "express": "^4.17.1",
-    "mongoose": "^5.9.15"
+    "mongoose": "^5.11.9"
   },
   "devDependencies": {
     "concurrently": "^5.3.0",
-    "nodemon": "^2.0.4"
+    "nodemon": "^2.0.6",
+    "prettier": "^2.2.1"
   },
   "engines": {
-    "node": "10.3.0",
-    "npm": "6.5.0"
+    "node": "14",
+    "npm": "6"
   }
 }
