@@ -1,19 +1,18 @@
-const assert = require('assert');
-const axios = require('axios');
-const cheerio = require('cheerio');
+import assert from 'assert';
+import axios from 'axios';
+import cheerio from 'cheerio';
 
-cheerio.prototype.justtext = function() {
+cheerio.prototype.justtext = function () {
   return this.clone().children().remove().end().text();
-}
+};
 
-cheerio.prototype.textArray = function() {
-  return this.map(function() {
+cheerio.prototype.textArray = function () {
+  return this.map(function () {
     return cheerio(this).text();
   }).toArray();
-}
+};
 
-
-async function getCodeforcesProblem ({ contest, problem }) {
+async function getCodeforcesProblem({ contest, problem }) {
   try {
     const url = `https://codeforces.com/contest/${contest}/problem/${problem}`;
     const response = await axios.get(url, { maxRedirects: 0 });
@@ -25,8 +24,18 @@ async function getCodeforcesProblem ({ contest, problem }) {
     let sampleTests = [];
     const tests = statement.find('.sample-test').children();
     for (let i = 0; i < tests.length; i += 2) {
-      const input = tests.eq(i).find('pre').html().replace(/<br>/g, '\n').replace(/&#xA0;/g, '\xa0');
-      const output = tests.eq(i + 1).find('pre').html().replace(/<br>/g, '\n').replace(/&#xA0;/g, '\xa0');
+      const input = tests
+        .eq(i)
+        .find('pre')
+        .html()
+        .replace(/<br>/g, '\n')
+        .replace(/&#xA0;/g, '\xa0');
+      const output = tests
+        .eq(i + 1)
+        .find('pre')
+        .html()
+        .replace(/<br>/g, '\n')
+        .replace(/&#xA0;/g, '\xa0');
       sampleTests.push({ input, output });
     }
 
@@ -44,37 +53,34 @@ async function getCodeforcesProblem ({ contest, problem }) {
         notes: statement.find('.note > p').textArray(),
       },
       link: url,
-      submitLink: `https://codeforces.com/contest/${contest}/submit/${problem}`
+      submitLink: `https://codeforces.com/contest/${contest}/submit/${problem}`,
     };
-  }
-  catch (err) {
+  } catch (err) {
+    console.log(err);
     throw new Error('Could not find or parse problem');
   }
 }
 
-function getHackerrankProblem ({ contest, problem }) {
+function getHackerrankProblem({ contest, problem }) {
   try {
-    const url = (contest
+    const url = contest
       ? `https://www.hackerrank.com/contests/${contest}/challenges/${problem}`
-      : `https://www.hackerrank.com/challenges/${problem}/problem`);
+      : `https://www.hackerrank.com/challenges/${problem}/problem`;
     throw new Error('Not implemented');
-  }
-  catch (err) {
+  } catch (err) {
     throw new Error('Could not find or parse problem');
   }
 }
 
-function scraper (type, pid) {
+function scraper(type, pid) {
   if (type === 'CF') {
     // Codeforces
     const match = pid.match(/([0-9]+)([A-Z][A-Z0-9]*)$/);
-    if (!match)
-      throw new Error('Invalid Codeforces problem ID');
+    if (!match) throw new Error('Invalid Codeforces problem ID');
     const contest = match[1];
     const problem = match[2];
     return getCodeforcesProblem({ contest, problem });
-  }
-  else if (type === 'HR') {
+  } else if (type === 'HR') {
     let contest = null;
     let problem = pid;
     const idx = pid.indexOf('/');
@@ -83,10 +89,9 @@ function scraper (type, pid) {
       problem = pid.substring(idx + 1);
     }
     return getHackerrankProblem({ contest, problem });
-  }
-  else {
+  } else {
     throw new Error('Invalid problem type');
   }
 }
 
-module.exports = scraper;
+export default scraper;
