@@ -1,10 +1,9 @@
-const crypto = require('crypto');
-const { promisify } = require('util');
+import crypto from 'crypto';
+import { promisify } from 'util';
+
 const randomBytesAsync = promisify(crypto.randomBytes);
 
-async function randomId(size = 6) {
+export async function randomId(size = 6) {
   const buf = await randomBytesAsync(size);
   return buf.toString('base64').replace(/\//g, '_').replace(/\+/g, '-');
 }
-
-module.exports = { randomId };
