@@ -1,9 +1,10 @@
-import React from 'react';
 import './Code.css';
 
 function Code({ value }) {
   return (
-    <pre className="code-element"><code>{value}</code></pre>
+    <pre className="code-element">
+      <code>{value}</code>
+    </pre>
   );
 }
 
