@@ -1,7 +1,5 @@
-import React from 'react';
-
 function Spacer({ width }) {
-  return <span style={{ marginLeft: (width || 'auto') }}></span>;
+  return <span style={{ marginLeft: width || 'auto' }}></span>;
 }
 
 export default Spacer;
